Climbing: refactoring filter and sort
