+++ src/components/FeaturePanel/Climbing/CragsInAreaSort/types.ts
@@ -1,0 +1,1 @@
+export type SortBy = 'default' | 'routes' | 'photos' | 'alphabetical';
