+++ src/components/FeaturePanel/CragsInArea.tsx
@@ -24,16 +24,12 @@ import { PROJECT_ID } from '../../services/project';
 import { getClickHandler } from './FeatureImages/Image/helpers';
 import { MemberItem } from './MemberFeatures/MemberItem';
 import { RouteDistribution } from './Climbing/RouteDistribution';
-import {
-  CragsInAreaSort,
-  useCragsInAreaSort,
-} from './Climbing/CragsInAreaSort';
-import {
-  CragsInAreaFilter,
-  filterCrag,
-  useCragsInAreaFilter,
-} from './Climbing/CragsInAreaFilter';
+import { CragsInAreaSort } from './Climbing/CragsInAreaSort/CragsInAreaSort';
+import { CragsInAreaFilter } from './Climbing/CragsInAreaFilter/CragsInAreaFilter';
 import { useUserSettingsContext } from '../utils/UserSettingsContext';
+import { filterCrag } from './Climbing/CragsInAreaFilter/utils/filterCrag';
+import { useCragsInAreaFilter } from './Climbing/CragsInAreaFilter/utils/useCragsInAreaFilter';
+import { useCragsInAreaSort } from './Climbing/CragsInAreaSort/utils/useCragsInAreaSort';
 
 const isOpenClimbing = PROJECT_ID === 'openclimbing';
